+++ zingolib/src/lightclient/test_features.rs
@@ -1,9 +1,8 @@
-use zcash_keys::address::Address;
-use zcash_primitives::transaction::components::amount::NonNegativeAmount;
+use zcash_client_backend::{PoolType, ShieldedProtocol};
 
 use crate::{
     error::ZingoLibError,
-    utils::{address_from_str, zatoshis_from_u64},
+    utils::conversion::{address_from_str, testing::receivers_from_send_inputs},
     wallet::Pool,
 };
 
@@ -22,6 +21,20 @@ impl LightClient {
         .map_err(ZingoLibError::CantReadWallet)
     }
 
+    /// Test only lightclient method for calling `do_propose` with primitive rust types
+    ///
+    /// # Panics
+    ///
+    /// Panics if the address, amount or memo conversion fails.
+    pub async fn do_propose_test_only(
+        &self,
+        address_amount_memo_tuples: Vec<(&str, u64, Option<&str>)>,
+    ) -> Result<String, String> {
+        let _receivers =
+            receivers_from_send_inputs(address_amount_memo_tuples, &self.config().chain);
+        unimplemented!()
+    }
+
     /// Test only lightclient method for calling `do_send` with primitive rust types
     ///
     /// # Panics
@@ -31,23 +44,8 @@ impl LightClient {
         &self,
         address_amount_memo_tuples: Vec<(&str, u64, Option<&str>)>,
     ) -> Result<String, String> {
-        let receivers: Vec<(Address, NonNegativeAmount, Option<MemoBytes>)> =
-            address_amount_memo_tuples
-                .into_iter()
-                .map(|(address, amount, memo)| {
-                    let address = address_from_str(address, &self.config().chain)
-                        .expect("should be a valid address");
-                    let amount = zatoshis_from_u64(amount)
-                        .expect("should be inside the range of valid zatoshis");
-                    let memo = memo.map(|memo| {
-                        crate::wallet::utils::interpret_memo_string(memo.to_string())
-                            .expect("should be able to interpret memo")
-                    });
-
-                    (address, amount, memo)
-                })
-                .collect();
-
+        let receivers =
+            receivers_from_send_inputs(address_amount_memo_tuples, &self.config().chain);
         self.do_send(receivers).await.map(|txid| txid.to_string())
     }
 
@@ -56,7 +54,6 @@ impl LightClient {
     /// # Panics
     ///
     /// Panics if the address conversion fails.
-    #[cfg(feature = "test-features")]
     pub async fn do_shield_test_only(
         &self,
         pools_to_shield: &[Pool],
@@ -68,5 +65,21 @@ impl LightClient {
         self.do_shield(pools_to_shield, address)
             .await
             .map(|txid| txid.to_string())
+    }
+
+    /// gets the first address that will allow a sender to send to a specific pool, as a string
+    pub async fn get_base_address(&self, pooltype: PoolType) -> String {
+        match pooltype {
+            PoolType::Transparent => self.do_addresses().await[0]["receivers"]["transparent"]
+                .clone()
+                .to_string(),
+            PoolType::Shielded(ShieldedProtocol::Sapling) => self.do_addresses().await[0]
+                ["receivers"]["sapling"]
+                .clone()
+                .to_string(),
+            PoolType::Shielded(ShieldedProtocol::Orchard) => {
+                self.do_addresses().await[0]["address"].take().to_string()
+            }
+        }
     }
 }